Change SF-DEFAULT to use SourceForge RSS API.
This is better than the previous (HTML) URL because it includes
releases from subdirectories.  The previous URL relied on the "Looking
for the latest version? ..." link which may not always point to the
actual latest version of anything and can handle only one release and
thus doesn't handle releases in subdirs at all.